Extract shared ranking regression utilities into helpers.py
Move duplicated code from both ranking regression test files into
helpers.py to reduce maintenance burden:

- BaselineStore class: encapsulates YAML baseline load/save/cache
- Shared functions: create_problem, result_to_config_tuple,
  transpose_key, compare_rankings, is_dtype_supported,
  load_problem_sizes (with stride parameter)

Both test files now import from helpers and only contain their
mode-specific config generation and test parametrization.

Verified: estimation (72 passed, 4:11) and simulation (40 passed, 2:23)
both pass unchanged.

Made-with: Cursor
